requirements.txt
#sudo apt-get install libmysqlclient-dev
mysql-python
simplejson
PyCrypto
Flask
behave
alembic